% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleGame.R
\name{assembleGame}
\alias{assembleGame}
\title{Assemble Play-by-Play Data Function}
\usage{
assembleGame(url)
}
\arguments{
\item{url}{a url of the following form: http://www.nhl.com/scores/htmlreports/year/PL02gamenumber.HTM}
}
\value{
a data frame containing all of the play-by-play from a specific NHL game
}
\description{
A function to scrape all of the event level data from a www.nhl.com/scores/htmlreports website.
}
\examples{
assembleGame("http://www.nhl.com/scores/htmlreports/20192020/PL020001.HTM")
}
