\name{fold.frames}
\alias{fold.frames}

\title{fold.frames}
\description{
  Collapse a series of data frames with the same columns into one data frame.
}
\usage{
  fold.frames(frame.list)
}
\arguments{
  \item{frame.list}{A list of data frames.}
}
\details{
  Most useful with compile.all.games().  
}
\value{
  A single data frame with the same columns as the originals. 
}
\author{A.C. Thomas <act@acthomas.ca>}

