% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20102011pbp.R
\name{get20102011pbp}
\alias{get20102011pbp}
\title{Assemble 2010-2011 NHL Play-by-Play Data Function}
\usage{
get20102011pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2010-2011 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2010-2011 NHL season.
}
\examples{
get20102011pbp()
}
