% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20112012pbp.R
\name{get20112012pbp}
\alias{get20112012pbp}
\title{Assemble 2011-2012 NHL Play-by-Play Data Function}
\usage{
get20112012pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2011-2012 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2011-2012 NHL season.
}
\examples{
get20112012pbp()
}
