% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20122013pbp.R
\name{get20122013pbp}
\alias{get20122013pbp}
\title{Assemble 2012-2013 NHL Play-by-Play Data Function}
\usage{
get20122013pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2012-2013 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2012-2013 NHL season.
}
\examples{
get20122013pbp()
}
