% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20132014pbp.R
\name{get20132014pbp}
\alias{get20132014pbp}
\title{Assemble 2013-2014 NHL Play-by-Play Data Function}
\usage{
get20132014pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2013-2014 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2013-2014 NHL season.
}
\examples{
get20132014pbp()
}
