% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20142015pbp.R
\name{get20142015pbp}
\alias{get20142015pbp}
\title{Assemble 2014-2015 NHL Play-by-Play Data Function}
\usage{
get20142015pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2014-2015 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2014-2015 NHL season.
}
\examples{
get20142015pbp()
}
