% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20152016pbp.R
\name{get20152016pbp}
\alias{get20152016pbp}
\title{Assemble 2015-2016 NHL Play-by-Play Data Function}
\usage{
get20152016pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2015-2016 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2015-2016 NHL season.
}
\examples{
get20152016pbp()
}
