% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20162017pbp.R
\name{get20162017pbp}
\alias{get20162017pbp}
\title{Assemble 2016-2017 NHL Play-by-Play Data Function}
\usage{
get20162017pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2016-2017 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2016-2017 NHL season.
}
\examples{
get20162017pbp()
}
