% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20172018pbp.R
\name{get20172018pbp}
\alias{get20172018pbp}
\title{Assemble 2017-2018 NHL Play-by-Play Data Function}
\usage{
get20172018pbp()
}
\value{
a data frame containing all of the play-by-play data from the 2017-2018 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2017-2018 NHL season.
}
\examples{
get20172018pbp()
}
