% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20182019Season.R
\name{get20182019Season}
\alias{get20182019Season}
\title{Assemble 2018-2019 NHL Play-by-Play Data Function}
\usage{
get20182019Season()
}
\value{
a data frame containing all of the play-by-play data from the 2018-2019 NHL season
}
\description{
Returns a dataframe containing the period, teams, strength, time, active skaters and decription of each event from the 2018-2019 NHL season.
}
\examples{
get20182019Season()
}
