% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get20192020Season.R
\name{get20192020Season}
\alias{get20192020Season}
\title{Assemble 2019-2020 NHL Play-by-Play Data Function}
\usage{
get20192020Season()
}
\value{
a data frame containing all of the play-by-play data from the 2019-2020 NHL season
}
\description{
Assemble 2019-2020 NHL Play-by-Play Data Function
}
\examples{
get20192020Season()
}
