% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_events.R
\name{get_game_events}
\alias{get_game_events}
\title{Game Events}
\usage{
get_game_events(game_id)
}
\arguments{
\item{game_id}{ID for single game}
}
\description{
Function that scrapes NHL API for single game event data.
}
