% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_info.R
\name{get_game_info}
\alias{get_game_info}
\title{Game Information}
\usage{
get_game_info(year_start, year_end)
}
\arguments{
\item{year_start}{First desired year to scrape from in YYYY format.}

\item{year_end}{Last desired year to scrape to in YYYY format.}
}
\description{
Function that scrapes the NHL API for game information between specified years.
}
