% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_roster.R
\name{get_game_roster}
\alias{get_game_roster}
\title{Game Roster}
\usage{
get_game_roster(game_id)
}
\arguments{
\item{game_id}{ID for single game}
}
\description{
Function that scrapes NHL API for single game roster data.
}
