% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_shifts.R
\name{get_game_shifts}
\alias{get_game_shifts}
\title{Game Shifts}
\usage{
get_game_shifts(game_id)
}
\arguments{
\item{game_id}{ID for single game}
}
\description{
Function that scrapes NHL API for single game event data.
}
