% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalie_stats.R
\name{get_goalie_stats}
\alias{get_goalie_stats}
\title{Goalie Stats}
\usage{
get_goalie_stats(year_start, year_end)
}
\arguments{
\item{year_start}{First desired year to scrape from in YYYY format.}

\item{year_end}{Last desired year to scrape to in YYYY format.}
}
\description{
Function that scrapes the NHL API for goalie statistics between specified years.
}
