% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_game_events.R
\name{get_multi_game_events}
\alias{get_multi_game_events}
\title{Multi Game Events}
\usage{
get_multi_game_events(game_ids)
}
\arguments{
\item{game_ids}{string of game IDs}
}
\description{
Function that scrapes the NHL API for event data for multiple games.
}
