% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_game_rosters.R
\name{get_multi_game_rosters}
\alias{get_multi_game_rosters}
\title{Multi Game Rosters}
\usage{
get_multi_game_rosters(game_ids)
}
\arguments{
\item{game_ids}{string of game IDs}
}
\description{
Function that scrapes the NHL API for roster data for multiple games.
}
