% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_game_shifts.R
\name{get_multi_game_shifts}
\alias{get_multi_game_shifts}
\title{Multi Game Shfits}
\usage{
get_multi_game_shifts(game_ids)
}
\arguments{
\item{game_ids}{string of game IDs}
}
\description{
Function that scrapes the NHL API for shift data for multiple games.
}
