% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skater_stats.R
\name{get_skater_stats}
\alias{get_skater_stats}
\title{Skater Statistics}
\usage{
get_skater_stats(year_start, year_end)
}
\arguments{
\item{year_start}{First desired year to scrape from in YYYY format.}

\item{year_end}{Last desired year to scrape to in YYYY format.}
}
\description{
Function that scrapes the NHL API for skater statistics between specified years.
}
