\name{player.summary}
\alias{player.summary}

\title{Player summaries}
\description{
  Summarizes the event count for each player in the data frame.
}
\usage{
  player.summary (grand.data, roster.unique)
}
\arguments{
  \item{grand.data}{A single data frame of all events to be considered.}
  \item{roster.unique}{A list of all player IDs to consider.}
}
\value{
	An array with event counts for each player in the data frame.
}
\author{A.C. Thomas <act@acthomas.ca>}

