\name{nhlscrapr-data}
\alias{quadsarray}

\title{nhlscrapr: Included Data Sets}
\description{
  Data sets included with the nhlscrapr package.
}
\usage{
  quadsarray
}
\format{
  quadsarray: a series of quadrilaterals that define distinct areas of
  the offensive zone.
}
\author{A.C. Thomas <act@acthomas.ca>}
\keyword{datasets}
